:truck: Move pipeline → connectivity_matrix in CPAC.connectome
[skip ci]
